% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_names.R
\name{common_names}
\alias{common_names}
\title{common_names}
\usage{
common_names(
  name1,
  name2,
  initials_only = FALSE,
  order = TRUE,
  threshold = 1,
  r_letters = c(K = "C", W = "V", Y = "I", Z = "S")
)
}
\arguments{
\item{name1}{vector of strings}

\item{name2}{vector of strings}

\item{order}{if TRUE, initials should be in the same ordre}
}
\value{
character vector of correspondances betweens initiales of the two names
}
\description{
Look for commons initials between two names
}
