% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_similar.R
\name{is_similar}
\alias{is_similar}
\title{is_similar}
\usage{
is_similar(
  name1,
  name2,
  threshold = 1,
  r_letters = c(K = "C", W = "V", Y = "I", Z = "S")
)
}
\arguments{
\item{name1}{vector of names (character type)}

\item{name2}{vector of names (character type)}

\item{threshold}{integer, max difference between two words (see namedist)}

\item{r_letters}{named vector of letter to replace (see replace_letters)}
}
\value{
TRUE or false based on the threshold
}
\description{
Get if two names are similar based on the distance between them.
If the difference is superior of the threshold, then there is no match
}
