% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_names.R
\name{match_names}
\alias{match_names}
\title{match_names}
\usage{
match_names(
  name,
  list_of_names,
  threshold = 1,
  last_name = TRUE,
  r_letters = c(K = "C", W = "V", Y = "I", Z = "S"),
  center_names_w = 0.1
)
}
\arguments{
\item{name}{name to match}

\item{list_of_names}{list of names}

\item{threshold}{threshold of distance}

\item{last_name}{TRUE if last name is in last position,
else, last name is in first position}

\item{r_letters}{letters to change}
}
\value{
list of matches
}
\description{
match names
}
