% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedist.R
\name{namedist}
\alias{namedist}
\title{namedist}
\usage{
namedist(name1, name2, r_letters = c(K = "C", W = "V", Y = "I", Z = "S"))
}
\arguments{
\item{name1}{string}

\item{name2}{string}

\item{r_letters}{vector of letters to be replaced if necessary}
}
\value{
distance between the two names as integer value
}
\description{
Apply stringdist function to two names in which the option
replace_letters is available
}
