% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_letters.R
\name{replace_letters}
\alias{replace_letters}
\title{replace_letters}
\usage{
replace_letters(word, r_letters = c(K = "C", W = "V", Y = "I", Z = "S"))
}
\arguments{
\item{word}{: string in which replace letters to normalize the name}

\item{r_letters}{: named character vector of letters to replace}
}
\value{
same word but with replaced letters
}
\description{
Replace selected letters by substitute
}
